CmdLine: Allow installation of files from CKAN files.
`ckan install -c YourMod.ckan`

This inserts the described mod into the registry, and then installs
as per normal. If that mod was indexed from upstream, the relevant
version (and only that version) will be replaced by the one
described in the file.

This does not roll-back the meta-data insertion if the user chooses
not to install the mod. As a dev tool option, I'm not stressed by
this.

Closes #128.