Don't register DLLs which are part of an installed mod.
The solution I've got walks through all installed mods every time
we find a DLL, making it O(N^2). That would be terrible, but DLL
scans are so rare, and the number of mods we have is so small, that
I'm considering it perfectly acceptable until someone complains
about the speed.

Chances are we may completely remove support for pre-CKAN DLLs
in the future, anyway.

Closes #141 after the user does a `ckan scan`.