Our relationship resolver no longer tries to install conflicting mods.
If the user asked for conflicting mods, an InconsistentKraken is thrown.

If mods we're installing or depend upon conflict with recommended or
suggested mods, then the lesser mods are removed from consideration.

I'm a terrible person for not having huge numbers of test cases here,
but I've been tested on huge amounts of CKAN metadata and discovering
*all* my inconsistencies, so I'm pretty sure this works better than
before. :)

Example test:

    ./ckan.exe install CustomBiomes-Data-RSS CustomBiomes-Data-Stock