RelationshipResolver now checks sanity before installing.
This was a surprisingly easy change, and it results in a really cool
list telling the user why what they've asked for isn't possible.

This also gives the Registry some nicer interfaces, eliminates duplicate
lines in our sanity checkout output, and expands the things the
InconsistentKraken can work with.

Fixes #332.