RelationshipResolver no longer suggests conflicting mods
If something would conflict with a mod we already have installed, or are
about to install, we remove it from consideration. While the previous
commit in this sequence ensures we never try to install conflicting
mods, this change ensures we provide a *shorter list* of
inconsistencies, so the user isn't scratching their head quite so much
wondering why a whole slew of things were going to be installed when
they conflict.

Fixes #332.