CKAN.dll: LatestAvailable now requires a version number
This means we can never go looking up mods which may be the wrong
version for our install by accident. It may also fix some edge cases
like #602 where we may show the wrong versions being available.

Most of the changes in this commit are to the test cases, which now
require the mandatory argument. :)

Tested that this closes #610 with our NEAR test example.