Change the glyph for "not compatible" to 'X'
The current marker character, ✗ (U+2717, BALLOT X) uses a glyph that
is not present in most fonts. Primarily, it is not present in any of
the fonts that are available by default to the Windows console
(Consolas, Lucida Console, Raster Fonts). Instead of the desired
character, the Windows console will display a '?' glyph, which is the
same glyph used for "Unkown", making these states indistinguishable.

I have changed the glyph to a standard X in order to be compatible with
a wider range of fonts in common use.