Kill all curlsharp threads on user cancellation.
For some reason, the cancel button *doesn't work*, but it has
the same buggy behaviour in 1.6.6, so this looks like it's an
existing bug, rather than a new one.

This code makes the curlsharp code consistent with the native
download code, so they should be reasonably bug-for-bug compatible,
making things easier to fix when we figure out what's wrong with
the download button.