Use original string when downloading using libcurl
For some reason *most* mods on KS are fine if the URL is presented as
"URL with spaces" or "URL%20with%20spaces". However some (R&S
CapsuleDyne in particualar) are not.

I have no idea why.

Luckily, the solution here is easy; we just make sure Curl always uses
the original download string.

This comes with a one line change, and a bunch of test improvements that
allowed me to reach the ability to make a one line change. :)