Curl is less picky about SSL certs.
Yeah, this sucks, with this change we're no longer doing a full
certificate check, which reduces the value of SSL.

On the other hand, this change means folks can download from sites
like ksp.sarbian.com that use certificates from authorities that
don't seem to be recognised by libcurl.

A better change in the future would be to provide some of our own
CAs as well, provide our own mirror that CKAN falls back to, or
something else. But for now, we're doing this awful thing. :/