Update Spec.md with RFC 2119 reference
This spec seems to use the RFC 2119 terminology for standard definitions of terms like "MUST", "SHOULD NOT", etc. -- so attach the commonly-used and recommended boilerplate for it.

Not certain if this is the best place in the file for it but here it is.  I also italicized the terms rather than using the all-caps used by the RFC to match the conventions used by the rest of the spec.

See https://www.ietf.org/rfc/rfc2119.txt