Start to rework relationship resolver.
Should be a no-op (expect for the new relationship type) for clients that use existing ctors
Add in a new ctor which does not run the relationship resolution and add a method to do so.
Add in a new Relationship type for installed modules.
Add in a way to override the installed modules list. Usful in cases in which mods are also going to be removed.