Setting `x_netkan_force_v` will force a 'v' onto versions.
`netkan.exe` maintains version strings exactly as we see them in
the release information. However mod authors sometimes switch between
adding a 'v' onto the front of their releases, and sometimes leaving it
off.

This change introduces a `x_netkan_force_v` field. If present and set
to 'true', this will prepend a 'v' to the version string of any indexed
mod if a 'v' is not already present.

This should solve our recent pain with InfernalRobotics,
RasterPropMonitor-Core, and countless other mods.

This requires no changes to end-user clients, and no changes to specs
(although I'll send in a separate PR that documents it as part of the
netkan pseudo-spec). Users of any CKAN version can enjoy us enforcing
version string consistency.

Comes with test cases. :)