Download default metadata in .tar.gz rather than .zip format
At some point, we gained the ability to read `.tar.gz` files, but the
client continued to use `.zip` to download metadata. This change flips
us to use tarballs, which are *an order of magnitude smaller* than the
.zip files. This can make a big difference, especially on slow
connections.

This change will affect both new users, and existing users if and only
if they're using the default repository settings.