Docker support added to CKAN.
With Docker, I can now build and run CKAN without needing to maintain a Mono installation.

If you have Docker Compose 1.6.2+ installed (and Docker Engine of course), you can too!

The docker-compose.yml file is written for Linux as that's what I use. If you're a Mac or Windows user, change the volume to reflect the location of your KSP install. I have not tested it, but it should work.

All CKAN commands should work.

To build the software:
$ docker-compose build ckan

To update all mods:
$ docker-compose run --rm ckan

To install a mod:
$ docker-compose run --rm ckan bash
root@<container>:/source# ckan install MechJeb2

The minor change to Cmdline/Main.cs was to disable the 'as root' warning while running in a Docker container.  It's standard practice to run processes as root in those containers, so the warning was just noisy.