Added Docker support.
A small change to Cmdline/Main.cs to check for Docker before requiring --asroot plus a Dockerfile to build the container and a docker-compose.yml to connect the KSP installation directory to the container.  As written, the file is written for Linux -- users on Mac and Windows will need to change the directory in that file.

To build the software:

$ docker-compose build ckan

To update all installed mods:

$ docker-compose run --rm ckan

To install a mod:

$ docker-compose run --rm ckan bash
root@<container>:/source# ckan install MechJeb2

All CKAN commands should work from that prompt.