Docker updates to support cake!
The new build system introduces some minor changes:
* New commands are required to build the software.
* The binaries are stored in different places.

The previous version of this Dockerfile was using the wrong binary!

As a bonus, the configuration setting has been abstracted out as
a build argument.  To build CKAN in the Debug config, use this command:

$ docker build --build-arg config=Debug -t ckan-debug .

To use this container, a command line like this will be required:

$ docker run --rm -v ${KSPDIR}:/kspdir ckan-debug install MechJeb

The default configuration setting is Release.