Quasi-freeform tag patch
This patch alters the structure of the "Sort column" schema to
remove the potential problems with ontology inherent to defining
"Categories", and redefines it into an array of strings called
"Tags" which may be chosen from a predefined list, or custom.

Custom tags are restricted to lowercase alphanumerics and hyphens.
A list of potential predefined tags may be found in prior commits
or in the original issue #1021 opened by Andy81le.