Add language selection dropdown to settings
Add a dropdown ComboBox for language selection in the settings.
The choosen language is saved in the Core's configuration (JSON/Win32Registry).
By default it is `null`, until either CKAN is started in a language that is supported by CKAN (e.g. via `LC_ALL` env var or simple OS settings),
or the user manually selects a language in the settings.
The supported languages are saved in `CKAN(-Core).Utilities.AvailableLanguages`.

Also added a check in `LoadOrCreateConfiguration()`, so that it'll create a new configuration if the file exists but is empty.