Create skip-releases option for NetKAN
With the --skip-releases <int> option set, NetKAN skips the given number of releases for inflation.
This happens in the transformers for all the kref sources, just before the --releases option is applied to take multiple releases.

This feature is useful for backfilling old releases of a mod, where you don't want to re-inflate the newer ones just to get to the older ones.