Don't offer mod upgrade if it would violate another mod's dependencies
In 'IRegistryQuerierHelpers.HasUpdate()' and 'ModuleInstaller.CanInstall()' we didn't pass the currently
installed module version as 'moduleToRemove', thus the RelationshipResolver couldn't detect if another
mod depending on a specific version of a mod would break.

This made the GUI offer upgrades that would fail during the actual installation, and the 'Versions' tab
listing versions as compatible that weren't.