Fixes #3402 don't refresh search every keypress
Rebuilding the mod list dataview every time a key was pressed takes a
non-trivial amount of time that makes the ui feel unresponsive and sluggish.

There already existed a workaround for this for linux and macos, but it was
still horribly slow on windows.

This change simply normalizes this behavior across the OSes so that there is
always a small delay from the last, non-enter, keypress to the start of the
refresh.